serialization: Drop rootfs.type (which had only one legal value)
With 0c52d5de (serialization: remove windows-specific layers+base
rootfs, 2016-08-26, #211) this field became a single-choice field.  We
could keep the field to make the current "layers" approach explicit in
the face of future rootfs types, but we'd have to solidify the wording
around it to cover what should happen when rootfs.type was not set, or
if rootfs.type was an unrecognized value.  It seems easier to remove
it and make the "layers" approach implicit default, so that's what
this commit does.  We can restore the field later if/when we have
another rootfs type without breaking forward compatibility.

Code translating between Docker and OCI image renderings can
drop/restore this field while it is translating Docker ↔ OCI media
types, so I don't expect compatibility issues from this change.

Signed-off-by: W. Trevor King <wking@tremily.us>