layer: Require ustar (IEEE Std 1003.1-2013)
The idea with a spec like this is to define behavior so that different
implementations can interoperate reliably.  When there is an interop
problem betweem implementations A and B, it should be clear from the
spec whether implementation A is broken, implementation B is broken,
or the spec is insufficiently clear.

For example, pax defines 'g' and 'x' typeflags [1] that aren't part of
the older ustar (originally defined in IEEE Std 1003.1-1988) [2].
Before this commit, if implementation A produced a layer with a 'g' or
'x' typeflag and implementation B died unpacking it, it was unclear
whose fault it was.  With this commit, it is clearly A's fault
(because it is using features not defined for ustar).

If implementation A had produces a layer with an '2' typeflag (which
ustar specifies for symlinks) and implementation B died unpacking it,
it is B's fault.

If implementation A had produces a layer with an 'S' typeflag (which
GNU uses for sparse files [3]) and implementation B died unpacking it,
it is neither party's fault, because ustar explicitly makes those
values implementation-defined.  Interop around them is up to out of
band communication between the layer author and layer consumer, and is
not covered by this spec.

The previous "File Types" section listed sockets, but the ustar spec
has:

  Attempts to archive a socket using ustar interchange format shall
  produce a diagnostic message.

And I see no socket entry in Go's set of typeflag constants [4], so
I'm not sure how they were supported before.

Go has supported pax since v1.1 [5,6], and pax lets you do things
(like having symlink targets longer than 100 characters).  But we're
avoiding requiring support for PAX because of name-recognition issues
[7].

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/pax.html#tag_20_92_13_02
[2]: https://github.com/libarchive/libarchive/wiki/ManPageTar5#POSIX_ustar_Archives
[3]: https://github.com/libarchive/libarchive/wiki/ManPageTar5#gnu-tar-archives
[4]: https://golang.org/pkg/archive/tar/#pkg-constants
[5]: https://codereview.appspot.com/6700047
[6]: https://github.com/golang/go/commit/106827990466b9246f9395233882b1e816df398a
[7]: https://github.com/opencontainers/image-spec/pull/342#issuecomment-249518142

Signed-off-by: W. Trevor King <wking@tremily.us>