descriptor: MUST URIs and SHOULD http(s) for descriptor's 'urls'
The in-flight #431 is using JSON Schema's uri format [1], which
requires RFC 3986.  This commit makes that an official spec
requirement by adding a MUST to the Markdown.  It also SHOULDs the
http(s) schemes, because they're widely supported and are deployed in
existing Docker images.  And Vincent and Antonio both like SHOULDing
them [2] ;).

[1]: https://tools.ietf.org/html/draft-wright-json-schema-validation-00#section-7.3.6
[2]: https://github.com/opencontainers/image-spec/issues/436

Signed-off-by: W. Trevor King <wking@tremily.us>