manifest: Require unpackers to create a config.json from the referenced config
The image-tools framework has attempted this since d3ffc1ce
(oci-image-tool: implement create-runtime-bundle, 2016-06-02, #114),
but we didn't specify whether the translation was required or what the
output of the translation should be.  That means downstream consumers
of an unpacked image couldn't be sure if they'd find a config.json or
(if they found one) which runtime-spec versions it would be compatible
with.  With this commit, the config output becomes specified, so
consumers can post-process their config.json and/or invoke a
runtime-spec 1.0.0-rc2-compatible runtime on it without worrying about
the presence or version of the unpacked config.json.

I've picked 1.0.0-rc2 as the most recent runtime-spec commit.  As the
runtime-spec moves forward with more RCs, I expect we'll want to bump
this to keep up.  Once runtime-spec hits 1.0, we can probably freeze
the target, since post 1.0 releases in runtime-spec's 1.x line are
unlikely to make translation from the config format easier, and any
1.x-compatible runtime will be able to handle 1.0 configs.

Signed-off-by: W. Trevor King <wking@tremily.us>