config: forbid snapshotting of data volumes
This is necessary in order to make sure that unpackers all have sane
behaviour when it comes to handling image repacking and layer
generation. Since data volumes are generally bindmounts from sources
external to the image, it is not a good idea to snapshot said data --
and thus we should forbid it.

Signed-off-by: Aleksa Sarai <asarai@suse.de>