vendoring: isolate the vendor directory
Due to the new vendoring logic in the golang compiler it can cause
issues for projects that import a package that has vendored a package
that is used locally. See https://github.com/containers/image/pull/223

This change moves the vendored sources to the package that uses them,
rather than for the whole project. Also is explictly is not vendoring
code repos from "github.com/opencontainers/". For now we'll consider
these non-remote. Though versioning may likely be future concern.

Fixes #527
Obsoletes #528

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>