image-layout: make index.json refnames unique
Without this change, a generic image handling tool would not be able to
handle all cases with duplicate org.opencontainers.ref.name annotations,
due to ambiguity about which descriptor is the "best match".

While it has been floated that since the annotation is not actually
mandatory implementations should have alternative methods of referencing
entries in an image-index, this is quite a frustrating issue for
implementations that just want to use org.opencontainers.ref.name in the
"common" case.

Signed-off-by: Aleksa Sarai <asarai@suse.de>