config: Use an array of strings in the ChainID definition
There's no need to couple this to DiffIDs and layer application, since
the implementation in identity/chainid.go is just operating on an
array of strings.  In the chainid.go implementation, those strings
happen to be digests, but nothing digest-specific is needed.

  $ cat a.go
  package main

  import (
    _ "crypto/sha256"
    "fmt"
    "github.com/opencontainers/go-digest"
    "github.com/opencontainers/image-spec/identity"
  )

  func main() {
    for _, chainID := range identity.ChainIDs([]digest.Digest{"a", "b", "c"}) {
      fmt.Println(chainID)
    }
  }
  $ go run a.go
  a
  sha256:c8687a08aa5d6ed2044328fa6a697ab8e96dc34291e8c2034ae8c38e6fcc6d65
  sha256:fd4283a6ddaa46181b8c3bd3f2497dd8f801e80b395b78bd6331197ba8db877d

which matches the example I'm adding with this commit.

Signed-off-by: W. Trevor King <wking@tremily.us>